Menu split into components; main React menu working
